Some refactoring
* Moved part of WangPainter back to WangBrush (BrushMode / TileMode).
  Tile mode could be re-introduced in WangPainter, if it is useful
  enough for a scripting API function.

* Merged the remainder of WangPainter with WangFiller, renaming the
  result to WangPainter. It did not seem like there was any advantage to
  separating these.

* Fixed leaking WangPainter instance in TileLayerWangEdit.
