+++ src/tiled/stampbrush.h
@@ -34,7 +34,6 @@ class WangSet;
 
 class MapDocument;
 class StampActions;
-class WangFiller;
 
 /**
  * Implements a tile brush that acts like a stamp. It is able to paint a block
