+++ src/tiled/wangbrush.h
@@ -1,6 +1,7 @@
 /*
  * wangbrush.h
  * Copyright 2017, Benjamin Trotter <bdtrotte@ucsc.edu>
+ * Copyright 2020, Thorbjørn Lindeijer <thorbjorn@lindeijer.nl>
  *
  * This file is part of Tiled.
  *
@@ -21,9 +22,8 @@
 #pragma once
 
 #include "abstracttiletool.h"
-#include "wangset.h"
-#include "wangfiller.h"
 #include "wangpainter.h"
+#include "wangset.h"
 
 namespace Tiled {
 
@@ -32,6 +32,13 @@ class WangBrush : public AbstractTileTool
     Q_OBJECT
 
 public:
+    enum BrushMode {
+        PaintCorner,
+        PaintEdge,
+        PaintEdgeAndCorner,
+        Idle // no valid color selected
+    };
+
     WangBrush(QObject *parent = nullptr);
     ~WangBrush() override;
 
@@ -74,6 +81,7 @@ private:
     void beginPaint();
     void doPaint(bool mergeable);
     void updateBrush();
+    void updateBrushAt(WangPainter &painter, QPoint pos);
 
     // The point painting happens around
     // In tile mode, this is that tile
@@ -85,9 +93,9 @@ private:
     QPoint mLineStartPos;
     WangId::Index mWangIndex = WangId::Top;
 
-    WangPainter *mWangPainter = nullptr;
     const WangSet *mWangSet = nullptr;
     int mCurrentColor = 0;
+    BrushMode mBrushMode = Idle;
     bool mIsTileMode = false;
     bool mRotationalSymmetry = false;
     bool mLineStartSet = false;
